Extract Feature subclass to its own file
