+++ packages/alfa-media/src/feature/feature.ts
@@ -4,14 +4,10 @@ import {
   Number,
   Percentage,
   String,
-  Token,
 } from "@siteimprove/alfa-css";
 import { Device } from "@siteimprove/alfa-device";
 import { Hash } from "@siteimprove/alfa-hash";
 import { None, Option } from "@siteimprove/alfa-option";
-import { Parser } from "@siteimprove/alfa-parser";
-import { Err, Ok, Result } from "@siteimprove/alfa-result";
-import { Slice } from "@siteimprove/alfa-slice";
 
 import * as json from "@siteimprove/alfa-json";
 
@@ -19,10 +15,6 @@ import { Media } from "../media";
 import { Resolver } from "../resolver";
 
 const { isLength } = Length;
-
-const { delimited, either, map, mapResult, option, pair, right } = Parser;
-
-const { isString } = String;
 
 /**
  * @see https://drafts.csswg.org/mediaqueries/#media-feature
@@ -60,13 +52,6 @@ export class Feature<
 
   public matches(device: Device): boolean {
     switch (this._name) {
-      // case "orientation":
-      //   return this._value.some(
-      //     (value) =>
-      //       value.type === "string" &&
-      //       value.value === device.viewport.orientation
-      //   );
-
       case "scripting":
         return device.scripting.enabled
           ? this._value.every(
@@ -157,136 +142,7 @@ export namespace Feature {
       | Percentage.JSON;
   }
 
-  type Features = {
-    orientation: [Orientation.Value, Orientation];
-  };
-
-  export function from<K extends keyof Features>(
-    name: K,
-    value: Option<Features[K][0]>
-  ): Result<Features[K][1], string>;
-
-  export function from(name: string, value: Option<Value>): Err<string>;
-
-  export function from(
-    name: string,
-    value: Option<Value>
-  ): Result<Feature, string> {
-    switch (name) {
-      case "orientation":
-        return value.every(Orientation.validate)
-          ? Ok.of(Orientation.myof(value))
-          : Err.of("Orientation must be landscape or portrait");
-      default:
-        return Err.of("Unknown feature");
-    }
-  }
-
   export function isFeature(value: unknown): value is Feature {
     return value instanceof Feature;
-  }
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-mf-name
-   */
-  const parseName = map(Token.parseIdent(), (ident) =>
-    ident.value.toLowerCase()
-  );
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-mf-value
-   */
-  const parseValue = either(
-    either(
-      map(Token.parseNumber(), (number) => Number.of(number.value)),
-      map(Token.parseIdent(), (ident) => String.of(ident.value.toLowerCase()))
-    ),
-    either(
-      map(
-        pair(
-          Token.parseNumber((number) => number.isInteger),
-          right(
-            delimited(option(Token.parseWhitespace), Token.parseDelim("/")),
-            Token.parseNumber((number) => number.isInteger)
-          )
-        ),
-        (result) => {
-          const [left, right] = result;
-
-          return Percentage.of(left.value / right.value);
-        }
-      ),
-      Length.parse
-    )
-  );
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-mf-plain
-   */
-  const parsePlain: Parser<Slice<Token>, Feature, string> = mapResult(
-    pair(
-      parseName,
-      right(
-        delimited(option(Token.parseWhitespace), Token.parseColon),
-        parseValue
-      )
-    ),
-    (result) => {
-      const [name, value] = result;
-
-      return name === "orientation"
-        ? Feature.from(name, Option.of(value))
-        : (Ok.of(Feature.of(name, Option.of(value))) as Result<
-            Feature,
-            string
-          >);
-    }
-  );
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-mf-boolean
-   */
-  const parseBoolean = map(parseName, (name) =>
-    Feature.of<string, never>(name)
-  );
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-feature
-   */
-  export const parse = delimited(
-    Token.parseOpenParenthesis,
-    delimited(option(Token.parseWhitespace), either(parsePlain, parseBoolean)),
-    Token.parseCloseParenthesis
-  );
-}
-
-class Orientation extends Feature<"orientation", Orientation.Value> {
-  public static myof(value: Option<Orientation.Value>): Orientation {
-    return new Orientation(value);
-  }
-
-  private constructor(value: Option<Orientation.Value>) {
-    super("orientation", value);
-  }
-
-  public static validate(value: Feature.Value): value is Orientation.Value {
-    return (
-      isString(value) &&
-      (value.value === "landscape" || value.value === "portrait")
-    );
-  }
-
-  public matches(device: Device): boolean {
-    return this._value.every(
-      (value) => value.value === device.viewport.orientation
-    );
-  }
-}
-
-namespace Orientation {
-  export type Value = String<"landscape"> | String<"portrait">;
-
-  export function isOrientation(value: unknown): value is Orientation {
-    return value instanceof Orientation;
   }
 }
